feat: gateway update profile

+++ app/internal/profiles/dto/request/filterUpdateRequestDto.go
@@ -3,6 +3,6 @@ package request
 type FilterUpdateRequestDto struct {
 	SessionId    string `json:"sessionId"`
 	SearchGender string `json:"searchGender"`
-	AgeFrom      byte   `json:"ageFrom"`
-	AgeTo        byte   `json:"ageTo"`
+	AgeFrom      uint64 `json:"ageFrom"`
+	AgeTo        uint64 `json:"ageTo"`
 }
