+++ app/internal/profiles/dto/request/filterUpdateRequestRepositoryDto.go
@@ -7,7 +7,7 @@ import (
 type FilterUpdateRequestRepositoryDto struct {
 	SessionId    string    `json:"sessionId"`
 	SearchGender string    `json:"searchGender"`
-	AgeFrom      byte      `json:"ageFrom"`
-	AgeTo        byte      `json:"ageTo"`
+	AgeFrom      uint64    `json:"ageFrom"`
+	AgeTo        uint64    `json:"ageTo"`
 	UpdatedAt    time.Time `json:"updatedAt"`
 }
