+++ app/internal/profiles/dto/request/profileAddRequestDto.go
@@ -25,8 +25,8 @@ type ProfileAddRequestDto struct {
 	TelegramQueryId         string                 `json:"telegramQueryId"`
 	Latitude                float64                `json:"latitude"`
 	Longitude               float64                `json:"longitude"`
-	AgeFrom                 byte                   `json:"ageFrom"`
-	AgeTo                   byte                   `json:"ageTo"`
+	AgeFrom                 uint64                 `json:"ageFrom"`
+	AgeTo                   uint64                 `json:"ageTo"`
 	Distance                float64                `json:"distance"`
 	Page                    uint64                 `json:"page"`
 	Size                    uint64                 `json:"size"`
