+++ app/internal/profiles/entity/filterEntity.go
@@ -7,8 +7,8 @@ type FilterEntity struct {
 	SessionId    string    `json:"sessionId"`
 	SearchGender string    `json:"searchGender"`
 	LookingFor   string    `json:"lookingFor"`
-	AgeFrom      byte      `json:"ageFrom"`
-	AgeTo        byte      `json:"ageTo"`
+	AgeFrom      uint64    `json:"ageFrom"`
+	AgeTo        uint64    `json:"ageTo"`
 	Distance     float64   `json:"distance"`
 	Page         uint64    `json:"page"`
 	Size         uint64    `json:"size"`
